ДЗ-5.1 Разделение кода на модули: Генератор данных и утилитарные функции
